Keep the native provider module to the provider itself
The module named after the alien provider also carried the factory for the
app's shared HTTP client, left over from when it was the Gem API layer.

The factory moves next to the rest of the app's injection, where iOS keeps
its session configuration too. Two modules that depended on the provider
module without importing anything from it drop the edge, and the app now
declares the provider it imports.
