+++ android/app/build.gradle.kts
@@ -202,6 +202,7 @@ android {
 
 dependencies {
     implementation(project(":ui"))
+    implementation(project(":data:services:native-provider"))
     implementation(project(":data:services:gemstone"))
     implementation(project(":data:coordinators"))
