+++ android/data/coordinators/build.gradle.kts
@@ -46,7 +46,6 @@ android {
 dependencies {
     implementation(project(":data:services:gemstone"))
     implementation(project(":data:services:store"))
-    api(project(":data:services:native-provider"))
 
     compileOnly(libs.compose.runtime.annotation)
