+++ android/data/services/native-provider/build.gradle.kts
@@ -3,7 +3,6 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinJvmCompile
 
 plugins {
     alias(libs.plugins.android.library)
-    id("com.google.devtools.ksp")
 }
 
 android {
@@ -49,11 +48,8 @@ android {
 }
 
 dependencies {
-    api(project(":gemcore"))
-    api(libs.okhttp)
-
-    implementation(libs.hilt.android)
-    ksp(libs.hilt.compiler)
+    implementation(project(":gemcore"))
+    implementation(libs.okhttp)
 
     implementation(libs.kotlinx.coroutines.android)
